requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
google-generativeai==0.8.3
spacy==3.8.2
tenacity==9.0.0
redis==5.0.1
requests==2.32.3
